+++ src/main/java/com/sandymandy/pleasurecraft/shader/GetShaderPreprocessor.java
@@ -1,9 +1,0 @@
-package com.sandymandy.pleasurecraft.shader;
-
-import org.jetbrains.annotations.Nullable;
-
-public interface GetShaderPreprocessor {
-    default @Nullable ShaderPreprocessor pleasurecraft$getShaderPreprocessor() {
-        return null;
-    }
-}